requirements.txt
flask
python-telegram-bot
python-telegram-bot[socks]
moviepy
opencv-python-headless
